introduce post->user relation
create and implement `user` column in `posts` table as foreign key (pointing to `users`.`id`)
make sure only authenticated users can edit their own posts
hide invalid actions on index page

+++ src/Controller/DefaultController.php
@@ -25,34 +25,47 @@ class DefaultController extends AbstractController
     public function postEdit($id) {
         # retrieve results
         $result = $this->getDbConnection()->query("SELECT * FROM `posts` WHERE `id`='$id';")->fetch(\PDO::FETCH_ASSOC);
-
+        if ($result) {
+            if ($result['user'] !== $this->getUserId()) {
+                throw new \Exception('Users are not allowed to edit foreign Posts', 403);
+            }
+        } else if ($id) {
+            throw new \Exception('This Post does not exist', 404);
+        }
         $this->display('post_update.html', ['result' => $result]);        
-
     }
     public function postSave($data) {
         # TODO: perform mysql query sanitation
+        $user = $this->getUser();
+        if (!$user) {
+            throw new \Exception('Only authenticated Users may create Posts', 401);
+        }
         if ($data['id']) {
             if (!is_numeric($data['id'])) {
                 return;
             }
             return $this->getDbConnection()->exec("
                 UPDATE `posts` SET 
-                `title`='$data[title]', 
+                `title`='$data[title]',
                 `author`='$data[author]',
                 `content`='$data[content]',
                 `updated_at`='".date('Y-m-d H:i:s')."'
                 WHERE 
-                `id`='$data[id]';
+                `id`='$data[id]' AND
+                `user`='$user->id'
+                ;
                 ");     
         } else {
             $sql = "INSERT INTO `posts` (
                 `title`, 
+                `user`,
                 `author`,
                 `content`,
                 `created_at`,
                 `updated_at`
                 ) VALUES (
                     '$data[title]',
+                    '$user->id',
                     '$data[author]',
                     '$data[content]',
                     '".date('Y-m-d H:i:s')."',
@@ -66,7 +79,11 @@ class DefaultController extends AbstractController
         if (!$id || !is_numeric($id)) {
             return;
         }
-        $result = $this->getDbConnection()->exec("DELETE FROM `posts` WHERE `id`=$id;");
+        $uid = $this->getUserId();
+        if (!$uid) {
+            throw new \Exception("This resource is only available for authenticated users", 401);
+        }
+        $result = $this->getDbConnection()->exec("DELETE FROM `posts` WHERE `id`='$id' AND `user`='$uid' LIMIT 1;");
         return $this->index();
     }
